+++ src/components/pages/add-revision.tsx
@@ -12,7 +12,8 @@ import { EditorPageData } from '@/fetcher/fetch-editor-data'
 import {
   RevisionAddMutationData,
   useRevisionAddMutation,
-} from '@/helper/mutations/revision'
+} from '@/helper/mutations/use-revision-add-mutation'
+import { EdtrState } from '@/schema/edtr-io-types'
 
 export function AddRevision({
   initialState,
@@ -100,7 +101,8 @@ export function AddRevision({
                   // @ts-expect-error temporary
                   __typename: type === 'GroupedExercise' ? 'Exercise' : type,
                 },
-                _needsReview
+                _needsReview,
+                initialState as EdtrState
               )
               return new Promise((resolve, reject) => {
                 if (success) resolve()
