+++ src/components/user-tools/author-tools.tsx
@@ -14,8 +14,8 @@ import { MenuSubButtonLink } from './menu-sub-button-link'
 import { useCanDo } from '@/auth/use-can-do'
 import { useInstanceData } from '@/contexts/instance-context'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
-import { useSubscriptionSetMutation } from '@/helper/mutations/subscription'
-import { useSetUuidStateMutation } from '@/helper/mutations/uuid'
+import { useSetUuidStateMutation } from '@/helper/mutations/use-set-uuid-state-mutation'
+import { useSubscriptionSetMutation } from '@/helper/mutations/use-subscription-set-mutation'
 import { useIsSubscribed } from '@/helper/use-is-subscribed'
 
 export enum Tool {
